requirements.txt
googlesearch-python==1.0.1
cloudscraper
requests